+++ app/src/main/java/com/gemini/energy/data/local/AuditLocalDataSource.kt
@@ -6,7 +6,7 @@ import io.reactivex.Observable
 
 class AuditLocalDataSource(private val auditDao: AuditDao) {
 
-    fun get(id: Int): Observable<AuditLocalModel> = auditDao.get(id).toObservable()
+    fun get(id: Long): Observable<AuditLocalModel> = auditDao.get(id).toObservable()
     fun getAll(): Observable<List<AuditLocalModel>> = auditDao.getAll().toObservable()
 
     fun save(audit: AuditLocalModel): Observable<Unit> = Observable.fromCallable {
@@ -14,7 +14,7 @@ class AuditLocalDataSource(private val auditDao: AuditDao) {
     }
 
     fun update(audit: AuditLocalModel): Observable<Unit> = Observable.fromCallable { auditDao.update(audit) }
-    fun delete(auditId: Int): Observable<Unit> = Observable.fromCallable {
+    fun delete(auditId: Long): Observable<Unit> = Observable.fromCallable {
         auditDao.delete(auditId)
     }
