+++ app/src/main/java/com/gemini/energy/data/local/ZoneLocalDataSource.kt
@@ -7,7 +7,7 @@ import io.reactivex.Observable
 class ZoneLocalDataSource(private val zoneDao: ZoneDao) {
 
     fun get(id: Int): Observable<ZoneLocalModel> = zoneDao.get(id).toObservable()
-    fun getAllByAudit(id: Int): Observable<List<ZoneLocalModel>> = zoneDao.getAllByAudit(id).toObservable()
+    fun getAllByAudit(id: Long): Observable<List<ZoneLocalModel>> = zoneDao.getAllByAudit(id).toObservable()
 
     fun save(zone: ZoneLocalModel): Observable<Unit> = Observable.fromCallable {
         zoneDao.insert(zone)
@@ -16,7 +16,7 @@ class ZoneLocalDataSource(private val zoneDao: ZoneDao) {
     fun update(zone: ZoneLocalModel): Observable<Unit> = Observable.fromCallable { zoneDao.update(zone) }
     fun delete(id: Int): Observable<Unit> = Observable.fromCallable { zoneDao.delete(id) }
 
-    fun deleteByAuditId(id: Int): Observable<Unit> = Observable.fromCallable {
+    fun deleteByAuditId(id: Long): Observable<Unit> = Observable.fromCallable {
         zoneDao.deleteByAuditId(id)
     }
