+++ app/src/main/java/com/gemini/energy/data/local/model/GraveLocalModel.kt
@@ -7,6 +7,6 @@ import android.arch.persistence.room.PrimaryKey
 class GraveLocalModel(
         @PrimaryKey
         var usn: Int,
-        var oid: Int,
+        var oid: Long,
         var type: Int
 )
