+++ app/src/main/java/com/gemini/energy/data/repository/TypeRepository.kt
@@ -15,7 +15,7 @@ class TypeRepository(
     fun get(id: Int): Observable<TypeLocalModel> = typeLocalDataSource.get(id)
     fun getAllTypeByZone(id: Int, type: String): Observable<List<TypeLocalModel>> =
             typeLocalDataSource.getAllTypeByZone(id, type)
-    fun getAllTypeByAudit(id: Int): Observable<List<TypeLocalModel>> =
+    fun getAllTypeByAudit(id: Long): Observable<List<TypeLocalModel>> =
             typeLocalDataSource.getAllTypeByAudit(id)
 
     fun save(auditScope: Type): Observable<Unit> =
@@ -24,6 +24,6 @@ class TypeRepository(
     fun update(type: Type): Observable<Unit> = typeLocalDataSource.update(auditScopeMapper.toLocal(type))
     fun delete(id: Int): Observable<Unit> = typeLocalDataSource.delete(id)
     fun deleteByZoneId(id: Int): Observable<Unit> = typeLocalDataSource.deleteByZoneId(id)
-    fun deleteByAuditId(id: Int): Observable<Unit> = typeLocalDataSource.deleteByAuditId(id)
+    fun deleteByAuditId(id: Long): Observable<Unit> = typeLocalDataSource.deleteByAuditId(id)
 
 }
