+++ app/src/main/java/com/gemini/energy/domain/entity/Type.kt
@@ -9,7 +9,7 @@ data class Type(
     var subType: String?,
 
     val zoneId: Int?,
-    val auditId: Int?,
+    val auditId: Long?,
 
     val createdAt: Date?,
     val updatedAt: Date?
