+++ app/src/main/java/com/gemini/energy/domain/interactor/AuditDeleteUseCase.kt
@@ -7,9 +7,9 @@ import io.reactivex.Observable
 
 
 class AuditDeleteUseCase(schedulers: Schedulers, private val auditGateway: AuditGateway)
-    : UseCase<Int, Unit>(schedulers) {
+    : UseCase<Long, Unit>(schedulers) {
 
-    override fun buildObservable(params: Int?): Observable<Unit> {
+    override fun buildObservable(params: Long?): Observable<Unit> {
         return auditGateway.deleteAudit(params!!)
     }
