+++ app/src/main/java/com/gemini/energy/domain/interactor/AuditGetUseCase.kt
@@ -7,9 +7,9 @@ import com.gemini.energy.domain.gateway.AuditGateway
 import io.reactivex.Observable
 
 class AuditGetUseCase(schedulers: Schedulers, private val auditGateway: AuditGateway):
-        UseCase<Int, Audit>(schedulers) {
+        UseCase<Long, Audit>(schedulers) {
 
-    override fun buildObservable(params: Int?): Observable<Audit> {
+    override fun buildObservable(params: Long?): Observable<Audit> {
         return auditGateway.getAudit(params!!)
     }
