+++ app/src/main/java/com/gemini/energy/domain/interactor/FeatureGetAllUseCase.kt
@@ -7,9 +7,9 @@ import com.gemini.energy.domain.gateway.AuditGateway
 import io.reactivex.Observable
 
 class FeatureGetAllUseCase(schedulers: Schedulers, private val auditGateway: AuditGateway) :
-        UseCase<Int, List<Feature>>(schedulers) {
+        UseCase<Long, List<Feature>>(schedulers) {
 
-    override fun buildObservable(auditId: Int?): Observable<List<Feature>> {
+    override fun buildObservable(auditId: Long?): Observable<List<Feature>> {
         return auditGateway.getFeature(auditId!!)
     }
