+++ app/src/main/java/com/gemini/energy/presentation/audit/detail/zone/list/model/ZoneModel.kt
@@ -7,5 +7,5 @@ import kotlinx.android.parcel.Parcelize
 data class ZoneModel(
         val id: Int?,
         val name: String,
-        val auditId: Int
+        val auditId: Long
 ) : Parcelable
