+++ app/src/main/java/com/gemini/energy/presentation/type/list/model/TypeModel.kt
@@ -11,6 +11,6 @@ data class TypeModel(
         val subType: String?,
 
         val zoneId: Int?,
-        val auditId: Int?
+        val auditId: Long?
 ) : Parcelable
